Move java_version_parser to module utils
