+++ src/modules/mod.rs
@@ -12,7 +12,6 @@ mod git_status;
 mod golang;
 mod hostname;
 mod java;
-mod java_version_parser;
 mod jobs;
 mod kubernetes;
 mod line_break;
@@ -25,6 +24,7 @@ mod ruby;
 mod rust;
 mod time;
 mod username;
+mod utils;
 
 #[cfg(feature = "battery")]
 mod battery;
