+++ src/modules/utils/mod.rs
@@ -1,0 +1,1 @@
+pub mod java_version_parser;
